+++ scripts/validate.sh
@@ -1,10 +1,14 @@
 #!/usr/bin/env bash
 
+# Output colors
+NC='\033[0m' # No Color
+RED='\033[0;31m'
+GREEN='\033[0;32m'
+BLUE='\033[0;34m'
+
 # Ensures the previous command completed successfully otherwise prints given message and exits
 function validate {
   if [[ $? -ne 0 ]]; then
-    RED='\033[0;31m'
-    NC='\033[0m' # No Color
     # Use argument as message or default message if no argument was given
     msg=${1:-Something went wrong, scroll up for details.}
     echo -e "\n${RED}FAILED:" $msg"${NC}\n"